extract appcard component
